使用 BitBlt 替换 PrintWindow, 并且设置窗口置顶
原因在这里:

This is not a Windows API bug, they are bugs or incomplete implementations in the apps that own the window you're trying to print.  PrintWindow sends the window the WM_PRINT message and a device context.  The app that owns the window is then supposed to render the window contents to the device context.  Support for this is spotty, especially if it was done by Adobe programmers.

The alternative is to BitBlt the window contents off the screen.  To deal with layered windows, you need to turn on the CAPTUREBLT bit in the dwRop argument.  Check this thread for a reference implementation using managed code.

https://social.msdn.microsoft.com/Forums/en-US/d9a188ae-3503-4c31-aa00-6e7195b5384a/printwindow-bugs?forum=winforms